pyproject.toml
[project]
name = "test-plan"
version = "1.0.0"
description = "Claude Code skills for generating test plans and test cases"
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
    "pymarkdownlnt",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]